ncs36510: timer.h rename
This fixes an issue as Timer.h is mbed file, thus if this file gets included first,
causes failures.
