+++ targets/TARGET_ONSEMI/TARGET_NCS36510/ncs36510_us_ticker_api.c
@@ -28,7 +28,7 @@
 */
 
 #include <stddef.h>
-#include "timer.h"
+#include "timer_ncs36510.h"
 
 #define US_TIMER  TIMER0
 #define US_TICKER TIMER1
