+++ targets/TARGET_ONSEMI/TARGET_NCS36510/timer_ncs36510.h
